Move code around to be functionally together
