+++ src/main/java/org/opentestsystem/ap/imrt/iis/batch/impl/SpringBatchJobManagementServiceImpl.java
@@ -1,9 +1,9 @@
-package org.opentestsystem.ap.imrt.iis.service.impl;
+package org.opentestsystem.ap.imrt.iis.batch.impl;
 
 import org.opentestsystem.ap.imrt.common.exception.NotFoundException;
 import org.opentestsystem.ap.imrt.iis.exception.JobAbandonmentException;
 import org.opentestsystem.ap.imrt.iis.model.JobExecutionResponse;
-import org.opentestsystem.ap.imrt.iis.service.SpringBatchJobManagementService;
+import org.opentestsystem.ap.imrt.iis.batch.SpringBatchJobManagementService;
 import org.springframework.batch.core.JobExecution;
 import org.springframework.batch.core.JobExecutionException;
 import org.springframework.batch.core.JobInstance;
