--- src/main/java/org/opentestsystem/ap/imrt/iis/business/ItemRevisionComparisonRule.java
@@ -1,20 +1,0 @@
-package org.opentestsystem.ap.imrt.iis.business;
-
-import org.opentestsystem.ap.common.model.Item;
-import org.opentestsystem.ap.imrt.common.model.ItemGitInformation;
-import org.opentestsystem.ap.imrt.iis.model.ItemRevision;
-
-/**
- * A business rule to be applied as an item revision is being ingested
- */
-public interface ItemRevisionComparisonRule {
-    /**
-     * Compare the existing item and the new version of the item capturing any necessary information based on the item change.
-     *
-     * @param existingGitInformation The {@link ItemGitInformation} for the existing item
-     * @param existingItem the existing {@link Item}
-     * @param current  The new {@link ItemRevision} of the item being ingested
-     * @return The current item revision, modified by the rule as applicable
-     */
-    ItemRevision applyBusinessRule(ItemGitInformation existingGitInformation, Item existingItem, ItemRevision current);
-}
