+++ src/main/java/org/opentestsystem/ap/imrt/iis/controller/ItemIngestJobController.java
@@ -5,7 +5,7 @@ import org.opentestsystem.ap.imrt.iis.model.JobExecutionResponse;
 import org.opentestsystem.ap.imrt.iis.service.ItemDeleteNotificationHandler;
 import org.opentestsystem.ap.imrt.iis.service.ItemMigrationService;
 import org.opentestsystem.ap.imrt.iis.service.ItemSynchronizationService;
-import org.opentestsystem.ap.imrt.iis.service.SpringBatchJobManagementService;
+import org.opentestsystem.ap.imrt.iis.batch.SpringBatchJobManagementService;
 import org.springframework.batch.core.Job;
 import org.springframework.batch.core.JobExecution;
 import org.springframework.batch.core.JobParameters;
