+++ src/main/java/org/opentestsystem/ap/imrt/iis/core/ItemBankMapper.java
@@ -1,4 +1,4 @@
-package org.opentestsystem.ap.imrt.iis.service;
+package org.opentestsystem.ap.imrt.iis.core;
 
 import org.opentestsystem.ap.imrt.common.model.ItemGitInformation;
 import org.opentestsystem.ap.imrt.iis.model.ItemBankItemRevision;
