+++ src/main/java/org/opentestsystem/ap/imrt/iis/service/ItemRevisionProcessor.java
@@ -2,7 +2,7 @@ package org.opentestsystem.ap.imrt.iis.service;
 
 import org.opentestsystem.ap.common.model.Item;
 import org.opentestsystem.ap.imrt.common.model.ItemGitInformation;
-import org.opentestsystem.ap.imrt.iis.business.ItemRevisionComparisonRule;
+import org.opentestsystem.ap.imrt.iis.core.ItemRevisionComparisonRule;
 import org.opentestsystem.ap.imrt.iis.model.ItemRevision;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
