--- src/main/java/org/opentestsystem/ap/imrt/iis/service/SpringBatchJobManagementService.java
@@ -1,19 +1,0 @@
-package org.opentestsystem.ap.imrt.iis.service;
-
-import org.opentestsystem.ap.imrt.iis.model.JobExecutionResponse;
-
-/**
- * A service for interacting with Spring Batch jobs.
- */
-public interface SpringBatchJobManagementService {
-    /**
-     * Mark a {@link org.springframework.batch.core.JobExecution} as
-     * {@link org.springframework.batch.core.BatchStatus#ABANDONED}.  If the job is in a running state, it will be
-     * stopped first.
-     *
-     * @param jobInstanceId The id of the {@link org.springframework.batch.core.JobInstance} that owns the
-     *                      {@link org.springframework.batch.core.JobExecution} to abandon.
-     * @return A {@link org.opentestsystem.ap.imrt.iis.model.JobExecutionResponse} indicating success/failure
-     */
-    JobExecutionResponse abandon(final long jobInstanceId);
-}
