+++ src/main/java/org/opentestsystem/ap/imrt/iis/service/impl/ItemIngestServiceImpl.java
@@ -8,7 +8,7 @@ import org.opentestsystem.ap.imrt.iis.exception.ItemIngestException;
 import org.opentestsystem.ap.imrt.iis.model.ItemBankItemRevision;
 import org.opentestsystem.ap.imrt.iis.model.ItemRevision;
 import org.opentestsystem.ap.imrt.iis.repository.ItemGitInformationRepository;
-import org.opentestsystem.ap.imrt.iis.service.ItemBankMapper;
+import org.opentestsystem.ap.imrt.iis.core.ItemBankMapper;
 import org.opentestsystem.ap.imrt.iis.service.ItemIngestService;
 import org.opentestsystem.ap.imrt.iis.service.ItemRevisionProcessor;
 import org.opentestsystem.ap.imrt.iis.service.ItemRevisionService;
