--- src/test/java/org/opentestsystem/ap/imrt/iis/business/TimeInWorkflowRuleTest.java
@@ -1,90 +1,0 @@
-package org.opentestsystem.ap.imrt.iis.business;
-
-import org.junit.Before;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.mockito.Mock;
-import org.mockito.runners.MockitoJUnitRunner;
-import org.opentestsystem.ap.common.model.EqItem;
-import org.opentestsystem.ap.common.model.Workflow;
-import org.opentestsystem.ap.imrt.common.model.BaseItem;
-import org.opentestsystem.ap.imrt.common.model.ItemGitInformation;
-import org.opentestsystem.ap.imrt.common.service.OperationalEventService;
-import org.opentestsystem.ap.imrt.iis.model.ItemRevision;
-
-import java.time.Instant;
-
-import static org.mockito.Matchers.any;
-import static org.mockito.Mockito.never;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
-
-@RunWith(MockitoJUnitRunner.class)
-public class TimeInWorkflowRuleTest {
-
-    @Mock
-    private ItemRevision revision;
-
-    @Mock
-    private BaseItem baseItem;
-
-    @Mock
-    private ItemGitInformation itemGitInformation;
-
-    @Mock
-    private ItemGitInformation existing;
-
-    @Mock
-    private BaseItem existingBaseItem;
-
-    @Mock
-    private OperationalEventService operationalEventService;
-
-    private TimeInWorkflowRule rule;
-
-    @Before
-    public void setup() {
-        rule = new TimeInWorkflowRule(operationalEventService);
-
-        when(revision.getItem()).thenReturn(baseItem);
-        when(revision.getItemGitInformation()).thenReturn(itemGitInformation);
-        when(itemGitInformation.getCurrentCommitDate()).thenReturn(Instant.ofEpochSecond(55));
-
-        when(existing.getItem()).thenReturn(existingBaseItem);
-    }
-
-    @Test
-    public void shouldUpdateNullExisting() {
-        rule.applyBusinessRule(null, null, revision);
-        verify(baseItem).setWorkflowStatusSetAt(itemGitInformation.getCurrentCommitDate());
-    }
-
-    @Test
-    public void shouldUpdateStatusChange() {
-        when(existingBaseItem.getWorkflowStatus()).thenReturn("status");
-        when(baseItem.getWorkflowStatus()).thenReturn("status1");
-
-        EqItem item = new EqItem("123");
-        Workflow wf = new Workflow();
-        wf.setWorkflowStatusCode("status");
-        item.setWorkflow(wf);
-
-        rule.applyBusinessRule(existing, item, revision);
-        verify(baseItem).setWorkflowStatusSetAt(itemGitInformation.getCurrentCommitDate());
-    }
-
-    @Test
-    public void shouldNotUpdate() {
-        when(existingBaseItem.getWorkflowStatus()).thenReturn("status");
-        when(baseItem.getWorkflowStatus()).thenReturn("status");
-
-        EqItem item = new EqItem("123");
-        Workflow wf = new Workflow();
-        wf.setWorkflowStatusCode("status");
-        item.setWorkflow(wf);
-
-        rule.applyBusinessRule(existing, item, revision);
-        verify(baseItem, never()).setWorkflowStatusSetAt(any());
-
-    }
-}
