+++ src/test/java/org/opentestsystem/ap/imrt/iis/service/ItemRevisionProcessorTest.java
@@ -6,7 +6,7 @@ import org.junit.runner.RunWith;
 import org.mockito.Mock;
 import org.mockito.runners.MockitoJUnitRunner;
 import org.opentestsystem.ap.imrt.common.model.ItemGitInformation;
-import org.opentestsystem.ap.imrt.iis.business.ItemRevisionComparisonRule;
+import org.opentestsystem.ap.imrt.iis.core.ItemRevisionComparisonRule;
 import org.opentestsystem.ap.imrt.iis.model.ItemRevision;
 
 import static org.mockito.Mockito.verify;
