+++ src/test/java/org/opentestsystem/ap/imrt/iis/service/impl/ItemBankMapperImplTest.java
@@ -17,6 +17,7 @@ import org.opentestsystem.ap.imrt.common.model.ImrtItem;
 import org.opentestsystem.ap.imrt.common.model.ItemGitInformation;
 import org.opentestsystem.ap.imrt.common.model.ItemLog;
 import org.opentestsystem.ap.imrt.iis.builder.TestItemBankItemRevisionBuilder;
+import org.opentestsystem.ap.imrt.iis.core.impl.ItemBankMapperImpl;
 import org.opentestsystem.ap.imrt.iis.model.ItemBankItemRevision;
 import org.opentestsystem.ap.imrt.iis.model.ItemRevision;
