+++ src/test/java/org/opentestsystem/ap/imrt/iis/service/impl/ItemIngestServiceImplTest.java
@@ -22,7 +22,7 @@ import org.opentestsystem.ap.imrt.iis.client.ItemBankClient;
 import org.opentestsystem.ap.imrt.iis.model.ItemBankItemRevision;
 import org.opentestsystem.ap.imrt.iis.model.ItemRevision;
 import org.opentestsystem.ap.imrt.iis.repository.ItemGitInformationRepository;
-import org.opentestsystem.ap.imrt.iis.service.ItemBankMapper;
+import org.opentestsystem.ap.imrt.iis.core.ItemBankMapper;
 import org.opentestsystem.ap.imrt.iis.service.ItemRevisionProcessor;
 import org.opentestsystem.ap.imrt.iis.service.ItemRevisionService;
 import org.opentestsystem.ap.imrt.iis.service.ItemWorkQueueService;
