+++ src/test/java/org/opentestsystem/ap/imrt/iis/service/impl/SpringBatchJobManagementServiceImplTest.java
@@ -6,8 +6,9 @@ import org.junit.runner.RunWith;
 import org.mockito.Mock;
 import org.mockito.runners.MockitoJUnitRunner;
 import org.opentestsystem.ap.imrt.common.exception.NotFoundException;
+import org.opentestsystem.ap.imrt.iis.batch.impl.SpringBatchJobManagementServiceImpl;
 import org.opentestsystem.ap.imrt.iis.model.JobExecutionResponse;
-import org.opentestsystem.ap.imrt.iis.service.SpringBatchJobManagementService;
+import org.opentestsystem.ap.imrt.iis.batch.SpringBatchJobManagementService;
 import org.springframework.batch.core.BatchStatus;
 import org.springframework.batch.core.JobExecution;
 import org.springframework.batch.core.JobExecutionException;
